requirements.txt
Django==3.2
django-materialize==1.0.1a2
pytz==2024.1
djangorestframework==3.14.0